+++ app/src/main/java/com/zskjprojectj/andoubusinessside/utils/ScreenUtil.java
@@ -25,8 +25,8 @@ public class ScreenUtil {
         }
     }
 
-    public static int px2sp(Context context, float pxValue) {
-        final float fontScale = context.getResources().getDisplayMetrics().scaledDensity;
+    public static int px2sp(Context context, double pxValue) {
+        final double fontScale = context.getResources().getDisplayMetrics().scaledDensity;
         return (int) (pxValue / fontScale + 0.5f);
     }
